BUG-2093 move YtlRerunPolicy into ytl package
